# target/jaxrs-analyzer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.0.1-SNAPSHOT",
        "title": "test"
    },
    "host": "example.com",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/": {
            "get": {
                "consumes": [
                ],
                "produces": [
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/create/{username}-{challName}-{challDescription}": {
            "post": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "challDescription",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "challName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}/CheckPoint/{cpID}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "cpId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}/Checkpoint": {
            "post": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "nameCP",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}/Segment": {
            "post": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "Longueur",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "cp1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "cp2",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "nameSegment",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}/Segment/{segmentID}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "segmentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/challenge/{name}/Segment/{segmentID}/Obstacle": {
            "post": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "segmentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "distance",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "etape",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "nameObstacle",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/signin": {
            "put": {
                "consumes": [
                ],
                "produces": [
                    "text/plain"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/signup/{username}-{password}": {
            "post": {
                "consumes": [
                ],
                "produces": [
                    "text/plain"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
    }
}